chore: sc

+++ script/HookMiner.sol
@@ -2,21 +2,26 @@
 pragma solidity ^0.8.21;
 
 /// @title HookMiner - a library for mining hook addresses
-/// @dev This library is intended for `forge test` environments. There may be gotchas when using salts in `forge script` or `forge create`
+/// @dev This library is intended for `forge test` environments. There may be gotchas when using salts in `forge script`
+/// or `forge create`
 library HookMiner {
     // mask to slice out the bottom 14 bit of the address
     uint160 constant FLAG_MASK = 0x3FFF;
 
     // Maximum number of iterations to find a salt, avoid infinite loops
-    uint256 constant MAX_LOOP = 100_000;
+    uint constant MAX_LOOP = 100_000;
 
     /// @notice Find a salt that produces a hook address with the desired `flags`
-    /// @param deployer The address that will deploy the hook. In `forge test`, this will be the test contract `address(this)` or the pranking address
-    ///                 In `forge script`, this should be `0x4e59b44847b379578588920cA78FbF26c0B4956C` (CREATE2 Deployer Proxy)
+    /// @param deployer The address that will deploy the hook. In `forge test`, this will be the test contract
+    /// `address(this)` or the pranking address
+    ///                 In `forge script`, this should be `0x4e59b44847b379578588920cA78FbF26c0B4956C` (CREATE2 Deployer
+    /// Proxy)
     /// @param flags The desired flags for the hook address
     /// @param creationCode The creation code of a hook contract. Example: `type(Counter).creationCode`
-    /// @param constructorArgs The encoded constructor arguments of a hook contract. Example: `abi.encode(address(manager))`
-    /// @return hookAddress salt and corresponding address that was found. The salt can be used in `new Hook{salt: salt}(<constructor arguments>)`
+    /// @param constructorArgs The encoded constructor arguments of a hook contract. Example:
+    /// `abi.encode(address(manager))`
+    /// @return hookAddress salt and corresponding address that was found. The salt can be used in `new Hook{salt:
+    /// salt}(<constructor arguments>)`
     function find(
         address deployer,
         uint160 flags,
@@ -24,18 +29,12 @@ library HookMiner {
         bytes memory constructorArgs
     ) internal view returns (address, bytes32) {
         address hookAddress;
-        bytes memory creationCodeWithArgs = abi.encodePacked(
-            creationCode,
-            constructorArgs
-        );
+        bytes memory creationCodeWithArgs = abi.encodePacked(creationCode, constructorArgs);
 
-        uint256 salt;
+        uint salt;
         for (salt; salt < MAX_LOOP; salt++) {
             hookAddress = computeAddress(deployer, salt, creationCodeWithArgs);
-            if (
-                uint160(hookAddress) & FLAG_MASK == flags &&
-                hookAddress.code.length == 0
-            ) {
+            if (uint160(hookAddress) & FLAG_MASK == flags && hookAddress.code.length == 0) {
                 return (hookAddress, bytes32(salt));
             }
         }
@@ -43,29 +42,18 @@ library HookMiner {
     }
 
     /// @notice Precompute a contract address deployed via CREATE2
-    /// @param deployer The address that will deploy the hook. In `forge test`, this will be the test contract `address(this)` or the pranking address
-    ///                 In `forge script`, this should be `0x4e59b44847b379578588920cA78FbF26c0B4956C` (CREATE2 Deployer Proxy)
+    /// @param deployer The address that will deploy the hook. In `forge test`, this will be the test contract
+    /// `address(this)` or the pranking address
+    ///                 In `forge script`, this should be `0x4e59b44847b379578588920cA78FbF26c0B4956C` (CREATE2 Deployer
+    /// Proxy)
     /// @param salt The salt used to deploy the hook
     /// @param creationCode The creation code of a hook contract
     function computeAddress(
         address deployer,
-        uint256 salt,
+        uint salt,
         bytes memory creationCode
     ) internal pure returns (address hookAddress) {
         return
-            address(
-                uint160(
-                    uint256(
-                        keccak256(
-                            abi.encodePacked(
-                                bytes1(0xFF),
-                                deployer,
-                                salt,
-                                keccak256(creationCode)
-                            )
-                        )
-                    )
-                )
-            );
+            address(uint160(uint(keccak256(abi.encodePacked(bytes1(0xFF), deployer, salt, keccak256(creationCode))))));
     }
 }
