+++ test/fees/HypeFeeCalculator.t.sol
@@ -1,17 +1,18 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.26;
 
-import {IPoolManager} from "@uniswap/v4-core/src/interfaces/IPoolManager.sol";
-import {IHooks} from "@uniswap/v4-core/src/libraries/Hooks.sol";
-import {TickMath} from "@uniswap/v4-core/src/libraries/TickMath.sol";
-import {toBalanceDelta} from "@uniswap/v4-core/src/types/BalanceDelta.sol";
-import {Currency} from "@uniswap/v4-core/src/types/Currency.sol";
-import {PoolKey} from "@uniswap/v4-core/src/types/PoolKey.sol";
-import {PoolId, PoolIdLibrary} from "@uniswap/v4-core/src/types/PoolId.sol";
+import {IPoolManager} from '@uniswap/v4-core/src/interfaces/IPoolManager.sol';
+import {IHooks} from '@uniswap/v4-core/src/libraries/Hooks.sol';
+import {TickMath} from '@uniswap/v4-core/src/libraries/TickMath.sol';
+import {toBalanceDelta} from '@uniswap/v4-core/src/types/BalanceDelta.sol';
+import {Currency} from '@uniswap/v4-core/src/types/Currency.sol';
 
-import {HypeFeeCalculator} from "@flaunch/fees/HypeFeeCalculator.sol";
-import {FairLaunch} from "@flaunch/hooks/FairLaunch.sol";
-import {FlaunchTest} from "../FlaunchTest.sol";
+import {PoolId, PoolIdLibrary} from '@uniswap/v4-core/src/types/PoolId.sol';
+import {PoolKey} from '@uniswap/v4-core/src/types/PoolKey.sol';
+
+import {FlaunchTest} from '../FlaunchTest.sol';
+import {HypeFeeCalculator} from '@flaunch/fees/HypeFeeCalculator.sol';
+import {FairLaunch} from '@flaunch/hooks/FairLaunch.sol';
 
 contract HypeFeeCalculatorTest is FlaunchTest {
     using PoolIdLibrary for PoolKey;
@@ -51,14 +52,15 @@ contract HypeFeeCalculatorTest is FlaunchTest {
     }
 
     /// setTargetTokensPerSec()
-    function test_CanSetTargetTokensPerSec(uint256 targetRate) public {
+    function test_CanSetTargetTokensPerSec(
+        uint targetRate
+    ) public {
         vm.assume(targetRate > 0);
 
         vm.prank(POSITION_MANAGER);
         feeCalculator.setFlaunchParams(poolId, abi.encode(targetRate));
 
-        (uint256 totalTokensSold, uint256 targetTokensPerSec) = feeCalculator
-            .poolInfos(poolId);
+        (uint totalTokensSold, uint targetTokensPerSec) = feeCalculator.poolInfos(poolId);
         assertEq(targetTokensPerSec, targetRate);
         assertEq(totalTokensSold, 0);
     }
@@ -85,32 +87,21 @@ contract HypeFeeCalculatorTest is FlaunchTest {
 
         // Mock fair launch not active
         vm.mockCall(
-            address(mockFairLaunch),
-            abi.encodeCall(mockFairLaunch.inFairLaunchWindow, poolId),
-            abi.encode(false)
+            address(mockFairLaunch), abi.encodeCall(mockFairLaunch.inFairLaunchWindow, poolId), abi.encode(false)
         );
 
-        assertEq(
-            feeCalculator.determineSwapFee(
-                poolKey,
-                _getSwapParams(1e18),
-                baseFee
-            ),
-            baseFee
-        );
+        assertEq(feeCalculator.determineSwapFee(poolKey, _getSwapParams(1e18), baseFee), baseFee);
     }
 
     function test_CalculatesHypeFee() public {
         uint24 baseFee = 100; // 1%
-        uint256 targetRate = 1000; // tokens per second
-        uint256 fairLaunchStart = block.timestamp;
-        uint256 fairLaunchEnd = fairLaunchStart + 30 minutes;
+        uint targetRate = 1000; // tokens per second
+        uint fairLaunchStart = block.timestamp;
+        uint fairLaunchEnd = fairLaunchStart + 30 minutes;
 
         // Setup fair launch window
         vm.mockCall(
-            address(mockFairLaunch),
-            abi.encodeCall(mockFairLaunch.inFairLaunchWindow, poolId),
-            abi.encode(true)
+            address(mockFairLaunch), abi.encodeCall(mockFairLaunch.inFairLaunchWindow, poolId), abi.encode(true)
         );
 
         vm.mockCall(
@@ -137,11 +128,7 @@ contract HypeFeeCalculatorTest is FlaunchTest {
         vm.prank(POSITION_MANAGER);
         _trackSwap(1000); // 1x target rate
         assertEq(
-            feeCalculator.determineSwapFee(
-                poolKey,
-                _getSwapParams(1e18),
-                baseFee
-            ),
+            feeCalculator.determineSwapFee(poolKey, _getSwapParams(1e18), baseFee),
             feeCalculator.MINIMUM_FEE_SCALED() / 100
         );
 
@@ -149,64 +136,42 @@ contract HypeFeeCalculatorTest is FlaunchTest {
         vm.warp(fairLaunchStart + 2);
         vm.prank(POSITION_MANAGER);
         _trackSwap(1500);
-        uint24 fee1 = feeCalculator.determineSwapFee(
-            poolKey,
-            _getSwapParams(1e18),
-            baseFee
-        );
+        uint24 fee1 = feeCalculator.determineSwapFee(poolKey, _getSwapParams(1e18), baseFee);
         assertEq(fee1, 13_25);
 
         // Third swap well above target - should get even higher fee
         vm.warp(fairLaunchStart + 3);
         vm.prank(POSITION_MANAGER);
         _trackSwap(3000);
-        uint24 fee2 = feeCalculator.determineSwapFee(
-            poolKey,
-            _getSwapParams(1e18),
-            baseFee
-        );
+        uint24 fee2 = feeCalculator.determineSwapFee(poolKey, _getSwapParams(1e18), baseFee);
         assertEq(fee2, 41_81);
 
         // Allow some time to pass - rate should decrease and fee should lower
         vm.warp(fairLaunchStart + 10);
-        uint24 fee3 = feeCalculator.determineSwapFee(
-            poolKey,
-            _getSwapParams(1e18),
-            baseFee
-        );
+        uint24 fee3 = feeCalculator.determineSwapFee(poolKey, _getSwapParams(1e18), baseFee);
         assertEq(fee3, 1_00);
 
         // Another large swap - fee should spike again
         vm.prank(POSITION_MANAGER);
         _trackSwap(5000);
-        uint24 fee4 = feeCalculator.determineSwapFee(
-            poolKey,
-            _getSwapParams(1e18),
-            baseFee
-        );
+        uint24 fee4 = feeCalculator.determineSwapFee(poolKey, _getSwapParams(1e18), baseFee);
         assertEq(fee4, 3_45);
 
         // Skip to near end of fair launch - rate and fee should be much lower
         vm.warp(fairLaunchEnd - 1 minutes);
-        uint24 fee5 = feeCalculator.determineSwapFee(
-            poolKey,
-            _getSwapParams(1e18),
-            baseFee
-        );
+        uint24 fee5 = feeCalculator.determineSwapFee(poolKey, _getSwapParams(1e18), baseFee);
         assertEq(fee5, 1_00);
     }
 
     function test_ReturnsMinimumFeeForLowRate() public {
         uint24 baseFee = 100; // 1%
-        uint256 targetRate = 1000; // tokens per second
-        uint256 fairLaunchStart = block.timestamp;
-        uint256 fairLaunchEnd = fairLaunchStart + 30 minutes;
+        uint targetRate = 1000; // tokens per second
+        uint fairLaunchStart = block.timestamp;
+        uint fairLaunchEnd = fairLaunchStart + 30 minutes;
 
         // Setup fair launch window
         vm.mockCall(
-            address(mockFairLaunch),
-            abi.encodeCall(mockFairLaunch.inFairLaunchWindow, poolId),
-            abi.encode(true)
+            address(mockFairLaunch), abi.encodeCall(mockFairLaunch.inFairLaunchWindow, poolId), abi.encode(true)
         );
 
         vm.mockCall(
@@ -234,27 +199,23 @@ contract HypeFeeCalculatorTest is FlaunchTest {
         _trackSwap(500); // 0.5x target rate
 
         // Fee should be minimum since rate is below target
-        uint24 fee = feeCalculator.determineSwapFee(
-            poolKey,
-            _getSwapParams(1e18),
-            baseFee
-        );
+        uint24 fee = feeCalculator.determineSwapFee(poolKey, _getSwapParams(1e18), baseFee);
         assertEq(fee, feeCalculator.MINIMUM_FEE_SCALED() / 100);
     }
 
-    function _trackSwap(int128 _amountSpecified) internal {
+    function _trackSwap(
+        int128 _amountSpecified
+    ) internal {
         feeCalculator.trackSwap(
             address(1),
             poolKey,
             IPoolManager.SwapParams({
                 zeroForOne: true,
                 amountSpecified: int(_amountSpecified),
-                sqrtPriceLimitX96: uint160(
-                    int160(TickMath.minUsableTick(poolKey.tickSpacing))
-                )
+                sqrtPriceLimitX96: uint160(int160(TickMath.minUsableTick(poolKey.tickSpacing)))
             }),
             toBalanceDelta(-(_amountSpecified / 2), _amountSpecified),
-            ""
+            ''
         );
     }
 }
